handle no trips or no charge, move utils.py to libs
