+++ web/figures.py
@@ -24,9 +24,7 @@ consumption_fig_by_speed = ERROR_DIV
 consumption_fig_by_temp = ERROR_DIV
 table_fig = ERROR_DIV
 info = ""
-battery_info = ERROR_DIV
-battery_table = None
-consumption_df_dict = None
+battery_table = ERROR_DIV
 
 AVG_CHARGE_SPEED = "avg_chg_speed"
 AVG_EMISSION_KM = "avg_emission_km"
@@ -52,8 +50,8 @@ SUMMARY_CARDS = {"Average consumption": {"text": [card_value_div(AVG_CONSUM_KW, 
 
 # pylint: disable=too-many-locals
 def get_figures(car: Car):
-    global consumption_fig, consumption_df, trips_map, consumption_fig_by_speed, table_fig, info, battery_info, \
-        battery_table, consumption_fig_by_temp, consumption_df_dict
+    global consumption_fig, consumption_df, trips_map, consumption_fig_by_speed, table_fig, info, \
+        battery_table, consumption_fig_by_temp
     lats = [42, 41]
     lons = [1, 2]
     names = ["undefined", "undefined"]
